+++ tests/smoke/CSharp50/AsyncAwait/AwaitTaskInsteadOfCallingTaskResult/TaskCanBeAwaitedInsteadOfCallingTaskResult.cs
@@ -1,6 +1,6 @@
 ﻿// ReSharper disable All
 
-// Expected number of suggestions: 3
+// Expected number of suggestions: 24
 
 using System.Threading.Tasks;
 
@@ -14,6 +14,68 @@ namespace CSharp50.AsyncAwait.AwaitTaskInsteadOfCallingTaskResult
             var result = task.Result;
             result = task.Result + 100;
             SomeMethod(task.Result);
+        }
+
+        async void CanBeAwaitedInMethodAndLocalMethod()
+        {
+            var task = new Task<int>(() => 0);
+            var result = task.Result;
+            result = task.Result + 100;
+            SomeMethod(task.Result);
+
+            async void LocalMethod()
+            {
+                task = new Task<int>(() => 0);
+                result = task.Result;
+                result = task.Result + 100;
+                SomeMethod(task.Result);
+            }
+        }
+
+        async void CanBeAwaitedInMethodAndLocalMethodAndNestedLocalMethod()
+        {
+            var task = new Task<int>(() => 0);
+            var result = task.Result;
+            result = task.Result + 100;
+            SomeMethod(task.Result);
+
+            async void LocalMethod()
+            {
+                task = new Task<int>(() => 0);
+                result = task.Result;
+                result = task.Result + 100;
+                SomeMethod(task.Result);
+
+                async void NestedLocalMethod()
+                {
+                    task = new Task<int>(() => 0);
+                    result = task.Result;
+                    result = task.Result + 100;
+                    SomeMethod(task.Result);
+                }
+            }
+        }
+
+        async void CanBeAwaitedWithinLocalMethodWithinAsyncMethod()
+        {
+            async void LocalMethod()
+            {
+                var task = new Task<int>(() => 0);
+                var result = task.Result;
+                result = task.Result + 100;
+                SomeMethod(task.Result);
+            }
+        }
+
+        void CanBeAwaitedWithinLocalMethodWithinNonAsyncMethod()
+        {
+            async void LocalMethod()
+            {
+                var task = new Task<int>(() => 0);
+                var result = task.Result;
+                result = task.Result + 100;
+                SomeMethod(task.Result);
+            }
         }
 
         void SomeMethod(int i) { }