+++ tests/smoke/CSharp50/AsyncAwait/AwaitTaskInsteadOfCallingTaskWait/TaskCanBeAwaitedInsteadOfCallingTaskWait.cs
@@ -1,6 +1,6 @@
 ﻿// ReSharper disable All
 
-// Expected number of suggestions: 3
+// Expected number of suggestions: 24
 
 using System.Threading.Tasks;
 
@@ -15,6 +15,75 @@ namespace CSharp50.AsyncAwait.AwaitTaskInsteadOfCallingTaskWait
             task.Wait(1 + 2 + 2);
             int wait = 100;
             task.Wait(wait);
+        }
+
+        async void CanBeAwaitedInMethodAndLocalMethod()
+        {
+            var task = new Task(() => { });
+            task.Wait(0);
+            task.Wait(1 + 2 + 2);
+            int wait = 100;
+            task.Wait(wait);
+
+            async void LocalMethod()
+            {
+                task = new Task(() => { });
+                task.Wait(0);
+                task.Wait(1 + 2 + 2);
+                wait = 100;
+                task.Wait(wait);
+            }
+        }
+
+        async void CanBeAwaitedInMethodAndLocalMethodAndNestedLocalMethod()
+        {
+            var task = new Task(() => { });
+            task.Wait(0);
+            task.Wait(1 + 2 + 2);
+            int wait = 100;
+            task.Wait(wait);
+
+            async void LocalMethod()
+            {
+                task = new Task(() => { });
+                task.Wait(0);
+                task.Wait(1 + 2 + 2);
+                wait = 100;
+                task.Wait(wait);
+
+                async void NestedLocalMethod()
+                {
+                    task = new Task(() => { });
+                    task.Wait(0);
+                    task.Wait(1 + 2 + 2);
+                    wait = 100;
+                    task.Wait(wait);
+                }
+            }
+        }
+
+        async void CanBeAwaitedWithinLocalMethodWithinAsyncMethod()
+        {
+            async void LocalMethod()
+            {
+                var task = new Task(() => { });
+                task.Wait(0);
+                task.Wait(1 + 2 + 2);
+                int wait = 100;
+                task.Wait(wait);
+            }
+        }
+
+        void CanBeAwaitedWithinLocalMethodWithinNonAsyncMethod()
+        {
+            async void LocalMethod()
+            {
+                var task = new Task(() => { });
+                task.Wait(0);
+                task.Wait(1 + 2 + 2);
+                int wait = 100;
+                task.Wait(wait);
+            }
         }
     }
 }