+++ tests/smoke/CSharp50/AsyncAwait/AwaitTaskWhenAllInsteadOfCallingTaskWaitAll/TaskWhenAllCanBeAwaitedInsteadOfCallingTaskWaitAll.cs
@@ -1,6 +1,6 @@
 ﻿// ReSharper disable All
 
-// Expected number of suggestions: 3
+// Expected number of suggestions: 24
 
 using System.Threading.Tasks;
 using static System.Threading.Tasks.Task;
@@ -14,6 +14,61 @@ namespace CSharp50.AsyncAwait.AwaitTaskWhenAllInsteadOfCallingTaskWaitAll
             Task.WaitAll(null);
             WaitAll(null);
             System.Threading.Tasks.Task.WaitAll(new Task[0]);
+        }
+
+        async void CanBeAwaitedInMethodAndLocalMethod()
+        {
+            Task.WaitAll(null);
+            WaitAll(null);
+            System.Threading.Tasks.Task.WaitAll(new Task[0]);
+
+            async void LocalMethod()
+            {
+                Task.WaitAll(null);
+                WaitAll(null);
+                System.Threading.Tasks.Task.WaitAll(new Task[0]);
+            }
+        }
+
+        async void CanBeAwaitedInMethodAndLocalMethodAndNestedLocalMethod()
+        {
+            Task.WaitAll(null);
+            WaitAll(null);
+            System.Threading.Tasks.Task.WaitAll(new Task[0]);
+
+            async void LocalMethod()
+            {
+                Task.WaitAll(null);
+                WaitAll(null);
+                System.Threading.Tasks.Task.WaitAll(new Task[0]);
+
+                async void NestedLocalMethod()
+                {
+                    Task.WaitAll(null);
+                    WaitAll(null);
+                    System.Threading.Tasks.Task.WaitAll(new Task[0]);
+                }
+            }
+        }
+
+        async void CanBeAwaitedWithinLocalMethodWithinAsyncMethod()
+        {
+            async void LocalMethod()
+            {
+                Task.WaitAll(null);
+                WaitAll(null);
+                System.Threading.Tasks.Task.WaitAll(new Task[0]);
+            }
+        }
+
+        void CanBeAwaitedWithinLocalMethodWithinNonAsyncMethod()
+        {
+            async void LocalMethod()
+            {
+                Task.WaitAll(null);
+                WaitAll(null);
+                System.Threading.Tasks.Task.WaitAll(new Task[0]);
+            }
         }
     }
 }