+++ tests/smoke/CSharp50/AsyncAwait/AwaitTaskWhenAllInsteadOfCallingTaskWaitAll/TaskWhenAllCannotBeAwaitedInsteadOfCallingTaskWaitAll.cs
@@ -1,5 +1,6 @@
 ﻿// ReSharper disable All
 
+using System;
 using System.Threading.Tasks;
 using static System.Threading.Tasks.Task;
 
@@ -11,6 +12,33 @@ namespace CSharp50.AsyncAwait.AwaitTaskWhenAllInsteadOfCallingTaskWaitAll
         {
             Task.WaitAll(null);
             WaitAll(null);
+        }
+
+        async void CannotBeAwaitedSoFarBecauseOfLambda()
+        {
+            Action a = () =>
+            {
+                Task.WaitAll(null);
+                WaitAll(null);
+            };
+        }
+
+        async void CannotBeAwaitedSoFarBecauseOfDelegate()
+        {
+            Action a = delegate ()
+            {
+                Task.WaitAll(null);
+                WaitAll(null);
+            };
+        }
+
+        async void CannotBeAwaitedBecauseNotDirectlyWithinTheCaller()
+        {
+            void LocalMethod()
+            {
+                Task.WaitAll(null);
+                WaitAll(null);
+            }
         }
     }
 }