+++ tests/smoke/CSharp50/AsyncAwait/AwaitTaskWhenAnyInsteadOfCallingTaskWaitAny/TaskWhenAnyCanBeAwaitedInsteadOfCallingTaskWaitAny.cs
@@ -1,6 +1,6 @@
 ﻿// ReSharper disable All
 
-// Expected number of suggestions: 4
+// Expected number of suggestions: 32
 
 using System.Threading.Tasks;
 using static System.Threading.Tasks.Task;
@@ -15,6 +15,68 @@ namespace CSharp50.AsyncAwait.AwaitTaskWhenAnyInsteadOfCallingTaskWaitAny
             WaitAny(null);
             System.Threading.Tasks.Task.WaitAny(new Task[0]);
             int value = WaitAny(null, 1000);
+        }
+
+        async void CanBeAwaitedInMethodAndLocalMethod()
+        {
+            Task.WaitAny(null);
+            WaitAny(null);
+            System.Threading.Tasks.Task.WaitAny(new Task[0]);
+            int value = WaitAny(null, 1000);
+
+            async void LocalMethod()
+            {
+                Task.WaitAny(null);
+                WaitAny(null);
+                System.Threading.Tasks.Task.WaitAny(new Task[0]);
+                value = WaitAny(null, 1000);
+            }
+        }
+
+        async void CanBeAwaitedInMethodAndLocalMethodAndNestedLocalMethod()
+        {
+            Task.WaitAny(null);
+            WaitAny(null);
+            System.Threading.Tasks.Task.WaitAny(new Task[0]);
+            int value = WaitAny(null, 1000);
+
+            async void LocalMethod()
+            {
+                Task.WaitAny(null);
+                WaitAny(null);
+                System.Threading.Tasks.Task.WaitAny(new Task[0]);
+                value = WaitAny(null, 1000);
+
+                async void NestedLocalMethod()
+                {
+                    Task.WaitAny(null);
+                    WaitAny(null);
+                    System.Threading.Tasks.Task.WaitAny(new Task[0]);
+                    value = WaitAny(null, 1000);
+                }
+            }
+        }
+
+        async void CanBeAwaitedWithinLocalMethodWithinAsyncMethod()
+        {
+            async void LocalMethod()
+            {
+                Task.WaitAny(null);
+                WaitAny(null);
+                System.Threading.Tasks.Task.WaitAny(new Task[0]);
+                int value = WaitAny(null, 1000);
+            }
+        }
+
+        void CanBeAwaitedWithinLocalMethodWithinNonAsyncMethod()
+        {
+            async void LocalMethod()
+            {
+                Task.WaitAny(null);
+                WaitAny(null);
+                System.Threading.Tasks.Task.WaitAny(new Task[0]);
+                int value = WaitAny(null, 1000);
+            }
         }
     }
 }