+++ tests/smoke/CSharp50/AsyncAwait/AwaitTaskWhenAnyInsteadOfCallingTaskWaitAny/TaskWhenAnyCannotBeAwaitedInsteadOfCallingTaskWaitAny.cs
@@ -1,5 +1,6 @@
 ﻿// ReSharper disable All
 
+using System;
 using System.Threading.Tasks;
 using static System.Threading.Tasks.Task;
 
@@ -11,6 +12,33 @@ namespace CSharp50.AsyncAwait.AwaitTaskWhenAnyInsteadOfCallingTaskWaitAny
         {
             Task.WaitAny(null);
             WaitAny(null);
+        }
+
+        async void CannotBeAwaitedSoFarBecauseOfLambda()
+        {
+            Action a = () =>
+            {
+                Task.WaitAny(null);
+                WaitAny(null);
+            };
+        }
+
+        async void CannotBeAwaitedSoFarBecauseOfDelegate()
+        {
+            Action a = delegate ()
+            {
+                Task.WaitAny(null);
+                WaitAny(null);
+            };
+        }
+
+        async void CannotBeAwaitedBecauseNotDirectlyWithinTheCaller()
+        {
+            void LocalMethod()
+            {
+                Task.WaitAny(null);
+                WaitAny(null);
+            }
         }
     }
 }